requirements.txt
requests==2.25.1
python-decouple==3.4

